File Moving

+++ src/sample/ide/fileTreeView/FileTreeView.java
@@ -3,24 +3,30 @@ package sample.ide.fileTreeView;
 import com.jfoenix.controls.JFXTreeView;
 import javafx.embed.swing.SwingFXUtils;
 import javafx.event.EventHandler;
+import javafx.scene.SnapshotParameters;
+import javafx.scene.control.MenuItem;
 import javafx.scene.control.*;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
-import javafx.scene.input.ContextMenuEvent;
+import javafx.scene.input.*;
+import javafx.scene.paint.Color;
 import org.apache.batik.transcoder.TranscoderException;
 import org.apache.batik.transcoder.TranscoderInput;
 import sample.ide.Ide;
 import sample.ide.tools.ComponentTabPane;
 
+import java.awt.*;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.HashMap;
+import java.nio.file.Files;
+import java.util.List;
+import java.util.*;
 
 public class FileTreeView extends JFXTreeView<File> {
 
     private static final HashMap<String, Image> IMAGE_CACHE = new HashMap<>();
+    private static final ArrayList<File> DELETE_ON_EXIT = new ArrayList<>();
     static {
         try {
             IMAGE_CACHE.put("default_file", imageFromSvg(FileTreeView.class.getResourceAsStream("icons/default_file.svg")));
@@ -55,46 +61,123 @@ public class FileTreeView extends JFXTreeView<File> {
 
     private static class CustomCell extends TreeCell<File> {
 
-//        private ComponentTabPane.ComponentTab<?> tab;
-        private final ContextMenu contextMenu = defaultContextMenu();
+        private ContextMenu contextMenu;
         private MenuItem requireIsFolder;
-        private final EventHandler<ContextMenuEvent> contextMenuEventEventHandler = contextMenuEvent -> {
-            if (contextMenuEvent.getSource() instanceof CustomCell) {
-                CustomCell customCell = (CustomCell) contextMenuEvent.getSource();
-                if (customCell.getItem() == null) {
-                    contextMenuEvent.consume();
-                } else  {
-                    requireIsFolder.setDisable(!customCell.getItem().isDirectory());
-                }
-            } else {
-                contextMenuEvent.consume();
-            }
-        };
+        private EventHandler<ContextMenuEvent> contextMenuEventEventHandler;
 
         public CustomCell() {
-            setOnMousePressed(mouseEvent -> {
-                if (mouseEvent.getClickCount() == 2) {
-                    if (getTreeItem() != null && ((CustomItem) getTreeItem()).getComponentTab() != null && getTreeView() != null && getTreeView() instanceof FileTreeView) {
-                        File value = getTreeItem().getValue();
-                        if (value != null && !value.isDirectory()) {
-                            ((CustomItem) getTreeItem()).setComponentTab(Ide.getNewEditorTab(value));
-                            TabPane tabPane = ((FileTreeView) getTreeView()).getIde().getTabPane();
-                            if (!tabPane.getTabs().contains(((CustomItem) getTreeItem()).getComponentTab())) {
-                                tabPane.getTabs().add(((CustomItem) getTreeItem()).getComponentTab());
-                            }
-                            tabPane.getSelectionModel().select(((CustomItem) getTreeItem()).getComponentTab());
-                        }
-                    } else if (getTreeItem() != null && ((CustomItem) getTreeItem()).getComponentTab() != null) {
-                        ((CustomItem) getTreeItem()).getComponentTab().getTabPane().getSelectionModel().select(((CustomItem) getTreeItem()).getComponentTab());
-                    }
-                }
-            });
-            setContextMenu(contextMenu);
-            setOnContextMenuRequested(contextMenuEventEventHandler);
             itemProperty().addListener((observableValue, treeItem, t1) -> {
                 if (getTreeView() != null && getTreeView() instanceof FileTreeView) {
                     ((FileTreeView) getTreeView()).getFileMap().remove(treeItem);
                     ((FileTreeView) getTreeView()).getFileMap().put(t1, this.getTreeItem());
+                }
+            });
+            treeItemProperty().addListener((observableValue, fileTreeItem, t1) -> {
+                if (contextMenu == null) {
+                    contextMenu = defaultContextMenu();
+                    setContextMenu(contextMenu);
+                }
+                if (contextMenuEventEventHandler == null) {
+                    contextMenuEventEventHandler = defaultContextMenuEvent();
+                    setOnContextMenuRequested(contextMenuEventEventHandler);
+                }
+                if (getOnMousePressed() == null) {
+                    setOnMousePressed(mouseEvent -> {
+                        if (mouseEvent.getClickCount() == 2 && getTreeItem() != null && getTreeItem() instanceof CustomItem) {
+                            CustomItem item = (CustomItem) getTreeItem();
+
+                            if (getTreeItem() != null && item.getComponentTab() == null && getTreeView() != null && getTreeView() instanceof FileTreeView) {
+                                File value = getTreeItem().getValue();
+                                if (value != null && !value.isDirectory()) {
+                                    item.setComponentTab(Ide.getNewEditorTab(value));
+                                    TabPane tabPane = ((FileTreeView) getTreeView()).getIde().getTabPane();
+                                    if (!tabPane.getTabs().contains(item.getComponentTab())) {
+                                        tabPane.getTabs().add(item.getComponentTab());
+                                    }
+                                    tabPane.getSelectionModel().select(item.getComponentTab());
+                                }
+                            } else if (getTreeItem() != null) {
+                                if (item.getComponentTab().getTabPane() != null) {
+                                    item.getComponentTab().getTabPane().getSelectionModel().select(item.getComponentTab());
+                                } else if (getTreeView() != null && getTreeView() instanceof FileTreeView) {
+                                    TabPane tabPane = ((FileTreeView) getTreeView()).getIde().getTabPane();
+                                    if (!tabPane.getTabs().contains(item.getComponentTab())) {
+                                        tabPane.getTabs().add(item.getComponentTab());
+                                    }
+                                    tabPane.getSelectionModel().select(item.getComponentTab());
+                                }
+                            }
+                        }
+                    });
+                }
+                if (getOnDragDetected() == null) {
+                    setOnDragDetected(mouseEvent -> {
+                        if (getItem() != null) {
+                            ClipboardContent clipboardContent = new ClipboardContent();
+                            ArrayList<File> arrayList = new ArrayList<>();
+                            arrayList.add(getItem());
+                            clipboardContent.putFiles(arrayList);
+                            Dragboard dragboard = this.startDragAndDrop(TransferMode.MOVE);
+                            dragboard.setContent(clipboardContent);
+                            SnapshotParameters snapshotParameters = new SnapshotParameters();
+                            snapshotParameters.setFill(Color.TRANSPARENT);
+                            dragboard.setDragView(this.snapshot(new SnapshotParameters(), null));
+                        }
+                    });
+                }
+                if (getOnDragOver() == null) {
+                    setOnDragOver(dragEvent -> {
+                        if (dragEvent.getDragboard().getContent(DataFormat.FILES) != null && getItem() != null && getItem().isDirectory()) {
+                            if (getItem().listFiles() != null) {
+                                List<File> files = (List<File>) dragEvent.getDragboard().getContent(DataFormat.FILES);
+                                if (!Arrays.asList(Objects.requireNonNull(getItem().listFiles())).contains(files.get(0))) {
+                                    dragEvent.acceptTransferModes(TransferMode.MOVE);
+                                    if (!getStyleClass().contains("file-drag-over")) {
+                                        getStyleClass().add("file-drag-over");
+                                    }
+                                }
+                            } else {
+                                if (!getStyleClass().contains("file-drag-over")) {
+                                    getStyleClass().add("file-drag-over");
+                                }
+                            }
+                        }
+                    });
+                }
+                if (getOnDragExited() == null) {
+                    setOnDragExited(dragEvent -> getStyleClass().remove("file-drag-over"));
+                }
+                if (getOnDragDropped() == null) {
+                    setOnDragDropped(dragEvent -> {
+                        if (dragEvent.getDragboard().getContent(DataFormat.FILES) != null && getItem() != null && getItem().isDirectory()) {
+                            File thisFile = getItem();
+                            if (dragEvent.getDragboard().getContent(DataFormat.FILES) instanceof List) {
+                                List<File> files = (List<File>) dragEvent.getDragboard().getContent(DataFormat.FILES);
+                                TreeItem<File> treeItem = getTreeItem();
+                                dragEvent.setDropCompleted(true);
+                                try {
+                                    for (File file : files) {
+                                        File newFile = new File(thisFile.getPath() + "\\" + file.getName());
+                                        if (!newFile.exists() && file.exists()) {
+                                            Files.move(file.toPath(), newFile.toPath());
+                                            if (treeItem.isExpanded()) {
+                                                treeItem.getChildren().add(new CustomItem(file));
+                                            }
+                                        }
+                                    }
+                                } catch (IOException e) {
+                                    e.printStackTrace();
+                                }
+                            }
+                        }
+                    });
+                }
+                if (getOnDragDone() == null) {
+                    setOnDragDone(dragEvent -> {
+                        if (!getItem().exists()) {
+                            getTreeItem().getParent().getChildren().remove(this.getTreeItem());
+                        }
+                    });
                 }
             });
         }
@@ -110,7 +193,7 @@ public class FileTreeView extends JFXTreeView<File> {
                     e.printStackTrace();
                 }
                 if (getTreeItem() != null && ((CustomItem) getTreeItem()).getComponentTab() != null) {
-                    ((CustomItem) getTreeItem()).getComponentTab().setText(file.getName());
+                    ((CustomItem) getTreeItem()).getComponentTab().getLabel().setText(file.getName());
                 }
             } else {
                 setText("");
@@ -123,18 +206,40 @@ public class FileTreeView extends JFXTreeView<File> {
             MenuItem deleteItem = new MenuItem("Delete");
             MenuItem newFile = new MenuItem("New File");
             MenuItem rename = new MenuItem("Rename");
-            contextMenu.getItems().addAll(newFile, rename, deleteItem);
+            MenuItem openInFileExplorer = new MenuItem("Open in File Explorer");
+            MenuItem run = new MenuItem("Run File");
+            contextMenu.getItems().addAll(newFile, rename, deleteItem, openInFileExplorer, run);
             requireIsFolder = newFile;
+            run.setOnAction(actionEvent -> {
+                if (getItem() != null) {
+                    try {
+                        Desktop.getDesktop().open(getItem());
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                    }
+                }
+            });
+            openInFileExplorer.setOnAction(actionEvent -> {
+                if (getItem() != null) {
+                    try {
+                        Runtime.getRuntime().exec("explorer.exe /select," + getItem().getPath());
+                    } catch (IOException e) {
+                        e.printStackTrace();
+                    }
+                }
+            });
             rename.setOnAction(actionEvent -> {
                 if (this.getTreeView() instanceof FileTreeView) {
                     FileTreeView fileTreeView = (FileTreeView) this.getTreeView();
                     fileTreeView.getIde().showPopupForText("", this.getItem().getName(), gotten -> {
                         if (!gotten.contains("/") && !gotten.contains("\\")) {
                             File renameTo = new File(this.getItem().getParentFile().getPath() + "\\" + gotten);
-                            this.getItem().renameTo(renameTo);
-                            this.setText(gotten);
-                            if (getTreeItem() != null && ((CustomItem) getTreeItem()).getComponentTab() != null) {
-                                ((CustomItem) getTreeItem()).getComponentTab().setText(gotten);
+                            if (!renameTo.exists()) {
+                                this.getItem().renameTo(renameTo);
+                                this.setText(gotten);
+                                if (getTreeItem() != null && ((CustomItem) getTreeItem()).getComponentTab() != null) {
+                                    ((CustomItem) getTreeItem()).getComponentTab().setTitle(gotten);
+                                }
                             }
                         }
                     });
@@ -163,14 +268,33 @@ public class FileTreeView extends JFXTreeView<File> {
                 if (this.getTreeView() instanceof FileTreeView) {
                     FileTreeView fileTreeView = (FileTreeView) this.getTreeView();
                     fileTreeView.getIde().showConfirmation("Are you sure you want to delete this file?\nThis cannot be undone.", gotten -> {
-                        this.getItem().delete();
+                        if (getTreeItem().getParent() != null) {
+                            getTreeItem().getParent().getChildren().remove(getTreeItem());
+                        }
                         if (getTreeItem() != null && ((CustomItem) getTreeItem()).getComponentTab() != null && ((CustomItem) getTreeItem()).getComponentTab().getTabPane() != null) {
+//                            ((CustomItem) getTreeItem()).getComponentTab().getFile().delete();
                             ((CustomItem) getTreeItem()).getComponentTab().getTabPane().getTabs().remove(((CustomItem) getTreeItem()).getComponentTab());
                         }
+                        DELETE_ON_EXIT.add(getItem());
+                        this.getItem().deleteOnExit();
                     });
                 }
             });
             return contextMenu;
+        }
+        private EventHandler<ContextMenuEvent> defaultContextMenuEvent() {
+            return contextMenuEvent -> {
+                if (contextMenuEvent.getSource() instanceof CustomCell) {
+                    CustomCell customCell = (CustomCell) contextMenuEvent.getSource();
+                    if (customCell.getItem() == null) {
+                        contextMenuEvent.consume();
+                    } else  {
+                        requireIsFolder.setDisable(!customCell.getItem().isDirectory());
+                    }
+                } else {
+                    contextMenuEvent.consume();
+                }
+            };
         }
 
     }
@@ -190,8 +314,10 @@ public class FileTreeView extends JFXTreeView<File> {
             }
             this.expandedProperty().addListener((observableValue, aBoolean, t1) -> {
                 if (t1) {
-                    File[] files = file.listFiles();
-                    if (files != null) {
+                    File[] filesArray = file.listFiles();
+                    if (filesArray != null) {
+                        List<File> files = new LinkedList<>(Arrays.asList(filesArray));
+                        files.removeAll(DELETE_ON_EXIT);
                         ArrayList<TreeItem<File>> newTreeItems = new ArrayList<>();
                         for (File loopFile : files) {
                             newTreeItems.add(new CustomItem(loopFile));
@@ -208,6 +334,7 @@ public class FileTreeView extends JFXTreeView<File> {
         public ComponentTabPane.ComponentTab<?> getComponentTab() {
             return componentTab;
         }
+
     }
 
     private static ImageView imageViewFromFile(File f) throws TranscoderException {
@@ -230,7 +357,6 @@ public class FileTreeView extends JFXTreeView<File> {
         ImageView imageView = new ImageView(image);
         imageView.setFitHeight(20);
         imageView.setPreserveRatio(true);
-        System.out.println(image.getUrl());
 
         return imageView;
     }
